- backup work in progress after major rewrite of internals

+++ distribution/libs/plugins/function.html_select_date.php
@@ -2,7 +2,7 @@
 
 /**
  * Smarty plugin
- * 
+ *
  * @package Smarty
  * @subpackage PluginsFunction
  */
@@ -17,11 +17,11 @@ require_once(SMARTY_PLUGINS_DIR . 'shared.make_timestamp.php');
 
 /**
  * Smarty {html_select_date} plugin
- * 
+ *
  * Type:     function<br>
  * Name:     html_select_date<br>
  * Purpose:  Prints the dropdowns for date selection.
- * 
+ *
  * ChangeLog:
  * <pre>
  *            - 1.0 initial release
@@ -37,21 +37,22 @@ require_once(SMARTY_PLUGINS_DIR . 'shared.make_timestamp.php');
  *              dropdown to include given date unless explicitly set (Monte)
  *            - 1.3.4 fix behaviour of 0000-00-00 00:00:00 dates to match that
  *              of 0000-00-00 dates (cybot, boots)
- *            - 2.0 complete rewrite for performance,  
+ *            - 2.0 complete rewrite for performance,
  *              added attributes month_names, *_id
  * </pre>
- * 
+ *
  * @link http://www.smarty.net/docs/en/language.function.html.select.date.tpl {html_select_date}
  *      (Smarty online manual)
  * @version 2.0
- * @author Andrei Zmievski 
- * @author Monte Ohrt <monte at ohrt dot com> 
+ * @author Andrei Zmievski
+ * @author Monte Ohrt <monte at ohrt dot com>
  * @author Rodney Rehm
- * @param array                    $params   parameters
- * @param Smarty_Internal_Template $template template object
- * @return string 
+ * @param array $params   parameters
+ * @param Smarty $template template object
+ * @return string
  */
-function smarty_function_html_select_date($params, $template) {
+function smarty_function_html_select_date($params, $template)
+{
     // generate timestamps used for month names only
     static $_month_timestamps = null;
     static $_current_year = null;
@@ -154,7 +155,7 @@ function smarty_function_html_select_date($params, $template) {
             case 'month_id':
             case 'day_id':
             case 'year_id':
-                $$_key = (string) $_value;
+                $$_key = (string)$_value;
                 break;
 
             case 'display_days':
@@ -162,7 +163,7 @@ function smarty_function_html_select_date($params, $template) {
             case 'display_years':
             case 'year_as_text':
             case 'reverse_years':
-                $$_key = (bool) $_value;
+                $$_key = (bool)$_value;
                 break;
 
             default:
@@ -212,13 +213,13 @@ function smarty_function_html_select_date($params, $template) {
         $key .= '_year';
         $t = $$key;
         if ($t === null) {
-            $$key = (int) $_current_year;
+            $$key = (int)$_current_year;
         } else if ($t[0] == '+') {
-            $$key = (int) ($_current_year + trim(substr($t, 1)));
+            $$key = (int)$_current_year + (int)trim(substr($t, 1));
         } else if ($t[0] == '-') {
-            $$key = (int) ($_current_year - trim(substr($t, 1)));
+            $$key = (int)$_current_year - (int)trim(substr($t, 1));
         } else {
-            $$key = (int) $$key;
+            $$key = (int)$$key;
         }
     }
 
@@ -247,8 +248,8 @@ function smarty_function_html_select_date($params, $template) {
             $_html_years = '<select name="' . $_name . '"';
             if ($year_id !== null || $all_id !== null) {
                 $_html_years .= ' id="' . smarty_function_escape_special_chars(
-                                $year_id !== null ? ( $year_id ? $year_id : $_name ) : ( $all_id ? ($all_id . $_name) : $_name )
-                        ) . '"';
+                    $year_id !== null ? ($year_id ? $year_id : $_name) : ($all_id ? ($all_id . $_name) : $_name)
+                ) . '"';
             }
             if ($year_size) {
                 $_html_years .= ' size="' . $year_size . '"';
@@ -256,14 +257,14 @@ function smarty_function_html_select_date($params, $template) {
             $_html_years .= $_extra . $extra_attrs . '>' . $option_separator;
 
             if (isset($year_empty) || isset($all_empty)) {
-                $_html_years .= '<option value="">' . ( isset($year_empty) ? $year_empty : $all_empty ) . '</option>' . $option_separator;
+                $_html_years .= '<option value="">' . (isset($year_empty) ? $year_empty : $all_empty) . '</option>' . $option_separator;
             }
 
             $op = $start_year > $end_year ? -1 : 1;
             for ($i = $start_year; $op > 0 ? $i <= $end_year : $i >= $end_year; $i += $op) {
                 $_html_years .= '<option value="' . $i . '"'
-                        . ($_year == $i ? ' selected="selected"' : '')
-                        . '>' . $i . '</option>' . $option_separator;
+                    . ($_year == $i ? ' selected="selected"' : '')
+                    . '>' . $i . '</option>' . $option_separator;
             }
 
             $_html_years .= '</select>';
@@ -285,8 +286,8 @@ function smarty_function_html_select_date($params, $template) {
         $_html_months = '<select name="' . $_name . '"';
         if ($month_id !== null || $all_id !== null) {
             $_html_months .= ' id="' . smarty_function_escape_special_chars(
-                            $month_id !== null ? ( $month_id ? $month_id : $_name ) : ( $all_id ? ($all_id . $_name) : $_name )
-                    ) . '"';
+                $month_id !== null ? ($month_id ? $month_id : $_name) : ($all_id ? ($all_id . $_name) : $_name)
+            ) . '"';
         }
         if ($month_size) {
             $_html_months .= ' size="' . $month_size . '"';
@@ -294,7 +295,7 @@ function smarty_function_html_select_date($params, $template) {
         $_html_months .= $_extra . $extra_attrs . '>' . $option_separator;
 
         if (isset($month_empty) || isset($all_empty)) {
-            $_html_months .= '<option value="">' . ( isset($month_empty) ? $month_empty : $all_empty ) . '</option>' . $option_separator;
+            $_html_months .= '<option value="">' . (isset($month_empty) ? $month_empty : $all_empty) . '</option>' . $option_separator;
         }
 
         for ($i = 1; $i <= 12; $i++) {
@@ -302,8 +303,8 @@ function smarty_function_html_select_date($params, $template) {
             $_text = isset($month_names) ? smarty_function_escape_special_chars($month_names[$i]) : ($month_format == "%m" ? $_val : strftime($month_format, $_month_timestamps[$i]));
             $_value = $month_value_format == "%m" ? $_val : strftime($month_value_format, $_month_timestamps[$i]);
             $_html_months .= '<option value="' . $_value . '"'
-                    . ($_val == $_month ? ' selected="selected"' : '')
-                    . '>' . $_text . '</option>' . $option_separator;
+                . ($_val == $_month ? ' selected="selected"' : '')
+                . '>' . $_text . '</option>' . $option_separator;
         }
 
         $_html_months .= '</select>';
@@ -324,8 +325,8 @@ function smarty_function_html_select_date($params, $template) {
         $_html_days = '<select name="' . $_name . '"';
         if ($day_id !== null || $all_id !== null) {
             $_html_days .= ' id="' . smarty_function_escape_special_chars(
-                            $day_id !== null ? ( $day_id ? $day_id : $_name ) : ( $all_id ? ($all_id . $_name) : $_name )
-                    ) . '"';
+                $day_id !== null ? ($day_id ? $day_id : $_name) : ($all_id ? ($all_id . $_name) : $_name)
+            ) . '"';
         }
         if ($day_size) {
             $_html_days .= ' size="' . $day_size . '"';
@@ -333,7 +334,7 @@ function smarty_function_html_select_date($params, $template) {
         $_html_days .= $_extra . $extra_attrs . '>' . $option_separator;
 
         if (isset($day_empty) || isset($all_empty)) {
-            $_html_days .= '<option value="">' . ( isset($day_empty) ? $day_empty : $all_empty ) . '</option>' . $option_separator;
+            $_html_days .= '<option value="">' . (isset($day_empty) ? $day_empty : $all_empty) . '</option>' . $option_separator;
         }
 
         for ($i = 1; $i <= 31; $i++) {
@@ -341,8 +342,8 @@ function smarty_function_html_select_date($params, $template) {
             $_text = $day_format == '%02d' ? $_val : sprintf($day_format, $i);
             $_value = $day_value_format == '%02d' ? $_val : sprintf($day_value_format, $i);
             $_html_days .= '<option value="' . $_value . '"'
-                    . ($_val == $_day ? ' selected="selected"' : '')
-                    . '>' . $_text . '</option>' . $option_separator;
+                . ($_val == $_day ? ' selected="selected"' : '')
+                . '>' . $_text . '</option>' . $option_separator;
         }
 
         $_html_days .= '</select>';
